requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
crontab==1.0.1
dj-database-url==2.1.0
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.14.0
environ==1.0
EZSheets==2023.8.11
frozenlist==1.4.1
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
html2text==2024.2.26
httplib2==0.22.0
idna==3.6
lxml==5.1.0
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplegmail==4.1.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telnet2==1.1.4
twilio==8.13.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.0
watch==0.2.7
watchdog==3.0.0
whitenoise==6.6.0
yarl==1.9.4
